+++ src/test/org/AdvancedJavaEindopdracht/configuration/data/GlobalSettings.csv
