+++ src/test/org/AdvancedJavaEindopdracht/resource/controller/PowerpointControllerTest.java
@@ -1,21 +1,19 @@
 package org.AdvancedJavaEindopdracht.resource.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.AdvancedJavaEindopdracht.resource.model.*;
+import org.AdvancedJavaEindopdracht.resource.model.Powerpoint;
+import org.AdvancedJavaEindopdracht.resource.model.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.web.SpringJUnitWebConfig;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.context.WebApplicationContext;
-
-import java.util.Date;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
@@ -24,7 +22,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @Transactional
 @ContextConfiguration(classes = org.AdvancedJavaEindopdracht.configuration.DatabaseConfigTest.class)
 @SpringJUnitWebConfig(classes = org.AdvancedJavaEindopdracht.configuration.DatabaseConfigTest.class)
-public class RoleControllerTest {
+public class PowerpointControllerTest {
     @Autowired
     private WebApplicationContext wac;
 
@@ -36,52 +34,50 @@ public class RoleControllerTest {
     }
 
     @Test
-    public void testGetRoles() throws Exception {
-        mockMvc.perform(get("/roles").contentType(MediaType.APPLICATION_JSON))
+    public void testGetPowerpoints() throws Exception {
+        mockMvc.perform(get("/powerpoint").contentType(MediaType.APPLICATION_JSON))
                 .andExpect(header().string(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.[0].role").value("user"))
-                .andExpect(jsonPath("$.[1].role").value("admin"));
+                .andExpect(jsonPath("$.[0].path").value("test"))
+                .andExpect(jsonPath("$.[0].user.id").value(1));
     }
 
     @Test
-    public void testGetRole() throws Exception {
-        mockMvc.perform(get("/roles/1"))
+    public void testGetPowerpoint() throws Exception {
+        mockMvc.perform(get("/powerpoint/1"))
+                .andExpect(header().string(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.role").value("user"));
+                .andExpect(jsonPath("$.path").value("test"))
+                .andExpect(jsonPath("$.user.id").value(1));
     }
 
     @Test
-    @DirtiesContext
-    public void testDeleteRoles() throws Exception {
-        mockMvc.perform(delete("/roles/1"))
+    public void testDeletePowerpoint() throws Exception {
+        mockMvc.perform(delete("/powerpoint/1"))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.role").value("user"));
+                .andExpect(jsonPath("$.path").value("test"))
+                .andExpect(jsonPath("$.user.id").value(1));
     }
 
     @Test
-    public void testPostRole() throws Exception {
-        Role role = new Role();
-        role.setRole("test");
+    public void testPostPowerpoint() throws Exception {
+        Powerpoint pp = new Powerpoint();
+        User user = new User();
 
-        mockMvc.perform(post("/roles")
-                        .content(new ObjectMapper().writeValueAsString(role))
+        user.setName("test");
+        user.setPassword("test");
+        user.setApproved(true);
+        user.setProfileImagePath("test");
+        user.setEmail("test");
+
+        pp.setUser(user);
+        pp.setPath("test");
+
+        mockMvc.perform(post("/powerpoint")
+                        .content(new ObjectMapper().writeValueAsString(pp))
                         .contentType("application/json"))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.role").value("test"));
-
-    }
-
-    @Test
-    public void testPutRoles() throws Exception {
-        Role role = new Role();
-        role.setRole("test");
-
-        mockMvc.perform(put("/roles/1")
-                        .content(new ObjectMapper().writeValueAsString(role))
-                        .contentType("application/json"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.role").value("test"));
+                .andExpect(jsonPath("$.path").value("test"));
 
     }
 }
