+++ src/test/org/AdvancedJavaEindopdracht/resource/controller/RoleControllerTest.java
@@ -1,8 +1,7 @@
 package org.AdvancedJavaEindopdracht.resource.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.AdvancedJavaEindopdracht.resource.model.Powerpoint;
-import org.AdvancedJavaEindopdracht.resource.model.User;
+import org.AdvancedJavaEindopdracht.resource.model.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,7 +22,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @Transactional
 @ContextConfiguration(classes = org.AdvancedJavaEindopdracht.configuration.DatabaseConfigTest.class)
 @SpringJUnitWebConfig(classes = org.AdvancedJavaEindopdracht.configuration.DatabaseConfigTest.class)
-public class PowerpointControllerTest {
+public class RoleControllerTest {
     @Autowired
     private WebApplicationContext wac;
 
@@ -35,50 +34,52 @@ public class PowerpointControllerTest {
     }
 
     @Test
-    public void testGetPowerpoints() throws Exception {
-        mockMvc.perform(get("/powerpoint").contentType(MediaType.APPLICATION_JSON))
+    public void testGetRoles() throws Exception {
+        mockMvc.perform(get("/roles").contentType(MediaType.APPLICATION_JSON))
                 .andExpect(header().string(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.[0].path").value("test"))
-                .andExpect(jsonPath("$.[0].user.id").value(1));
+                .andExpect(jsonPath("$.[0].role").value("user"))
+                .andExpect(jsonPath("$.[1].role").value("admin"));
     }
 
     @Test
-    public void testGetPowerpoint() throws Exception {
-        mockMvc.perform(get("/powerpoint/1"))
-                .andExpect(header().string(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE))
+    public void testGetRole() throws Exception {
+        mockMvc.perform(get("/roles/1"))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.path").value("test"))
-                .andExpect(jsonPath("$.user.id").value(1));
+                .andExpect(jsonPath("$.role").value("user"));
     }
 
     @Test
-    public void testDeletePowerpoint() throws Exception {
-        mockMvc.perform(delete("/powerpoint/1"))
+    @DirtiesContext
+    public void testDeleteRoles() throws Exception {
+        mockMvc.perform(delete("/roles/1"))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.path").value("test"))
-                .andExpect(jsonPath("$.user.id").value(1));
+                .andExpect(jsonPath("$.role").value("user"));
     }
 
     @Test
-    public void testPostPowerpoint() throws Exception {
-        Powerpoint pp = new Powerpoint();
-        User user = new User();
+    public void testPostRole() throws Exception {
+        Role role = new Role();
+        role.setRole("test");
 
-        user.setName("test");
-        user.setPassword("test");
-        user.setApproved(true);
-        user.setProfileImagePath("test");
-        user.setEmail("test");
-
-        pp.setUser(user);
-        pp.setPath("test");
-
-        mockMvc.perform(post("/powerpoint")
-                        .content(new ObjectMapper().writeValueAsString(pp))
+        mockMvc.perform(post("/roles")
+                        .content(new ObjectMapper().writeValueAsString(role))
                         .contentType("application/json"))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.path").value("test"));
+                .andExpect(jsonPath("$.role").value("test"));
+
+    }
+
+    @Test
+    public void testPutRoles() throws Exception {
+        Role role = new Role();
+        role.setRole("test");
+
+        mockMvc.perform(put("/roles/1")
+                        .content(new ObjectMapper().writeValueAsString(role))
+                        .contentType("application/json"))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.role").value("test"));
 
     }
 }
