+++ src/test/org/AdvancedJavaEindopdracht/resource/repository/EventRepositoryTest.java
@@ -1,10 +1,8 @@
 package org.AdvancedJavaEindopdracht.resource.repository;
 
 import org.AdvancedJavaEindopdracht.resource.model.event.Event;
-import org.AdvancedJavaEindopdracht.resource.model.event.EventDto;
 import org.AdvancedJavaEindopdracht.resource.model.event.content.Content;
 import org.AdvancedJavaEindopdracht.resource.model.event.content.contentType.ContentType;
-import org.AdvancedJavaEindopdracht.resource.service.EventService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
